Modulize functions and move to proper file location
